app/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.110.0
fastapi-cache2==0.2.1
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
idna==3.6
importlib_resources==6.3.1
joblib==1.3.2
multidict==6.0.5
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pendulum==3.0.0
psutil==5.9.8
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
pymystem3==0.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
redis==5.0.3
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
threadpoolctl==3.3.0
time-machine==2.14.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.28.0
yarl==1.9.4
zipp==3.18.1
